src/main/python/requirements.txt
pykka
kafka-python